+++ packages/store/src/storage.ts
@@ -6,7 +6,7 @@ import { PassThrough, Readable, Transform } from 'stream';
 import { pipeline } from 'stream/promises';
 import { default as URL } from 'url';
 
-import { hasProxyTo } from '@verdaccio/config';
+import { getProxiesForPackage, hasProxyTo } from '@verdaccio/config';
 import {
   API_ERROR,
   API_MESSAGE,
@@ -22,7 +22,7 @@ import {
   pluginUtils,
   searchUtils,
   tarballUtils,
-  validatioUtils,
+  validationUtils,
 } from '@verdaccio/core';
 import { asyncLoadPlugin } from '@verdaccio/loaders';
 import {
@@ -58,7 +58,7 @@ import {
   UnPublishManifest,
   Version,
 } from '@verdaccio/types';
-import { createTarballHash, isObject, normalizeContributors } from '@verdaccio/utils';
+import { createTarballHash, normalizeContributors } from '@verdaccio/utils';
 
 import {
   PublishOptions,
@@ -123,8 +123,8 @@ class Storage {
     debug('change existing package for package %o revision %o', name, revision);
     debug(`change manifest tags for %o revision %o`, name, revision);
     if (
-      !validatioUtils.isObject(metadata.versions) ||
-      !validatioUtils.isObject(metadata[DIST_TAGS])
+      !validationUtils.isObject(metadata.versions) ||
+      !validationUtils.isObject(metadata[DIST_TAGS])
     ) {
       debug(`change manifest bad data for %o`, name);
       throw errorUtils.getBadData();
@@ -192,7 +192,7 @@ class Storage {
     username: string
   ): Promise<Manifest> {
     debug('remove tarball %s for %s', filename, name);
-    assert(validatioUtils.validateName(filename));
+    assert(validationUtils.validateName(filename));
     const storage: pluginUtils.StorageHandler = this.getPrivatePackageStorage(name);
     if (!storage) {
       debug(`method not implemented for storage`);
@@ -316,10 +316,19 @@ class Storage {
             return;
           }
 
+          if (res.statusCode === HTTP_STATUS.UNAUTHORIZED) {
+            debug('remote stream response 401');
+            passThroughRemoteStream.emit(
+              'error',
+              errorUtils.getUnauthorized(errorUtils.API_ERROR.UNAUTHORIZED_ACCESS)
+            );
+            return;
+          }
+
           if (
             !(res.statusCode >= HTTP_STATUS.OK && res.statusCode < HTTP_STATUS.MULTIPLE_CHOICES)
           ) {
-            debug('remote stream response ok');
+            debug('remote stream response %o', res.statusCode);
             passThroughRemoteStream.emit(
               'error',
               errorUtils.getInternalError(`bad uplink status code: ${res.statusCode}`)
@@ -677,6 +686,7 @@ class Storage {
         (plugin: pluginUtils.ManifestFilter<Config>) => {
           return typeof plugin.filter_metadata !== 'undefined';
         },
+        false,
         this.config?.serverSettings?.pluginPrefix,
         PLUGIN_CATEGORY.FILTER
       );
@@ -707,7 +717,7 @@ class Storage {
     filename: string,
     { signal }: { signal: AbortSignal }
   ): Promise<Readable> {
-    assert(validatioUtils.validateName(filename));
+    assert(validationUtils.validateName(filename));
     const storage: pluginUtils.StorageHandler = this.getPrivatePackageStorage(pkgName);
     if (typeof storage === 'undefined') {
       return this.createFailureStreamResponseNext();
@@ -898,16 +908,17 @@ class Storage {
   private getUpLinkForDistFile(pkgName: string, distFile: DistFile): IProxy {
     let uplink: IProxy | null = null;
 
-    for (const uplinkId in this.uplinks) {
+    for (const uplinkName in this.uplinks) {
       // refer to https://github.com/verdaccio/verdaccio/issues/1642
-      if (hasProxyTo(pkgName, uplinkId, this.config.packages)) {
-        uplink = this.uplinks[uplinkId];
+      if (hasProxyTo(pkgName, uplinkName, this.config.packages)) {
+        uplink = this.uplinks[uplinkName];
       }
     }
 
     if (uplink == null) {
-      debug('upstream not found creating one for %o', pkgName);
+      debug('upstream not found, creating one for %o', pkgName);
       uplink = new ProxyStorage(
+        `verdaccio-${pkgName}`,
         {
           url: distFile.url,
           cache: true,
@@ -932,7 +943,7 @@ class Storage {
       });
     } else if (
       isPublishablePackage(manifest as Manifest) === false &&
-      validatioUtils.isObject((manifest as StarManifestBody).users)
+      validationUtils.isObject((manifest as StarManifestBody).users)
     ) {
       debug('update manifest star');
       // if user request to apply a star to the manifest
@@ -950,7 +961,7 @@ class Storage {
         ...options,
       });
       return API_MESSAGE.PKG_CHANGED;
-    } else if (validatioUtils.validatePublishSingleVersion(manifest)) {
+    } else if (validationUtils.validatePublishSingleVersion(manifest)) {
       // if continue, the version to be published does not exist
       // we create a new package
       debug('publish a new version');
@@ -969,7 +980,7 @@ class Storage {
         this.logger.error({ err }, 'notify batch service has failed: @{err.message}');
       }
       return message;
-    } else if (validatioUtils.validateUnPublishSingleVersion(manifest)) {
+    } else if (validationUtils.validateUnPublishSingleVersion(manifest)) {
       debug('unpublish a version');
 
       await this.unPublishAPackage(manifest as UnPublishManifest, {
@@ -1124,7 +1135,7 @@ class Storage {
       throw errorUtils.getNotFound();
     }
     const hasPackage = await storage.hasPackage(pkgName);
-    debug('has package %o for %o', pkgName, hasPackage);
+    debug('has package %o is %o', pkgName, hasPackage);
     return hasPackage;
   }
 
@@ -1144,7 +1155,7 @@ class Storage {
     const username = options.requestOptions.username;
     debug('publishing a new package for %o as %o', name, username);
     let successResponseMessage;
-    const manifest: Manifest = { ...validatioUtils.normalizeMetadata(body, name) };
+    const manifest: Manifest = { ...validationUtils.normalizeMetadata(body, name) };
     const { _attachments, versions } = manifest;
 
     // validation step, if _attachments is not an object throw error
@@ -1307,7 +1318,7 @@ class Storage {
     { signal }
   ): Promise<PassThrough> {
     debug(`add a tarball for %o`, pkgName);
-    assert(validatioUtils.validateName(filename));
+    assert(validationUtils.validateName(filename));
 
     const shaOneHash = createTarballHash();
     const transformHash = new Transform({
@@ -1422,9 +1433,9 @@ class Storage {
 
       // if uploaded tarball has a different shasum, it's very likely that we
       // have some kind of error
-      if (validatioUtils.isObject(metadata.dist) && _.isString(metadata.dist.tarball)) {
+      if (validationUtils.isObject(metadata.dist) && _.isString(metadata.dist.tarball)) {
         const tarball = tarballUtils.extractTarballFromUrl(metadata.dist.tarball);
-        if (validatioUtils.isObject(data._attachments[tarball])) {
+        if (validationUtils.isObject(data._attachments[tarball])) {
           if (
             _.isNil(data._attachments[tarball].shasum) === false &&
             _.isNil(metadata.dist.shasum) === false
@@ -1495,6 +1506,7 @@ class Storage {
     name: string,
     username: string | undefined
   ): Promise<void> {
+    debug('creating new package %o for user %o', name, username);
     const storage: pluginUtils.StorageHandler = this.getPrivatePackageStorage(name);
 
     if (!storage) {
@@ -1665,7 +1677,7 @@ class Storage {
       // etag??
     });
 
-    // if either local data and upstream data are empty, we throw an error
+    // if both local data and upstream data are empty, we throw an error
     if (!remoteManifest && _.isNull(data)) {
       throw errorUtils.getNotFound(`${API_ERROR.NOT_PACKAGE_UPLINK}: ${name}`);
       // if the remote manifest is empty, we return local data
@@ -1700,13 +1712,13 @@ class Storage {
       # one uplink setup
       proxy: npmjs
 
-    A package requires uplinks syncronization if enables the proxy section, uplinks
-    can be more than one, the more are the most slow request will take, the request
-    are made in serial and if 1st call fails, the second will be triggered, otherwise
+    A package requires uplinks syncronization if the proxy section is defined. There can be
+    more than one uplink. The more uplinks are defined, the longer the request will take.
+    The requests are made in serial and if 1st call fails, the second will be triggered, otherwise
     the 1st will reply and others will be discarded. The order is important.
 
-    Errors on upkinks are considered are, time outs, connection fails and http status 304,
-    in that case the request returns empty body and we want ask next on the list if has fresh
+    Errors on uplinks that are considered are time outs, connection fails, and http status 304.
+    In these cases the request returns empty body and we want ask next on the list if has fresh
     updates.
    */
   public async syncUplinksMetadata(
@@ -1716,20 +1728,18 @@ class Storage {
   ): Promise<[Manifest | null, any]> {
     let found = localManifest !== null;
     let syncManifest: Manifest | null = null;
-    const upLinks: string[] = [];
+    let upLinks: string[] = [];
     const hasToLookIntoUplinks = _.isNil(options.uplinksLook) || options.uplinksLook;
     debug('is sync uplink enabled %o', hasToLookIntoUplinks);
 
-    for (const uplink in this.uplinks) {
-      if (hasProxyTo(name, uplink, this.config.packages) && hasToLookIntoUplinks) {
-        debug('sync uplink %o', uplink);
-        upLinks.push(uplink);
-      }
+    if (hasToLookIntoUplinks) {
+      upLinks = getProxiesForPackage(name, this.config.packages);
+      debug('uplinks found for %o: %o', name, upLinks);
     }
 
-    //  if none uplink match we return the local manifest
+    //  if no uplinks match we return the local manifest
     if (upLinks.length === 0) {
-      debug('no uplinks found for %o upstream update aborted', name);
+      debug('no uplinks found for %o, upstream update aborted', name);
       return [localManifest, []];
     }
 
@@ -1774,9 +1784,9 @@ class Storage {
       debug('uplinks sync failed with %o errors', uplinksErrors.length);
       for (const err of uplinksErrors) {
         const { code } = err;
-        if (code === 'ETIMEDOUT' || code === 'ESOCKETTIMEDOUT' || code === 'ECONNRESET') {
+        if (code === HTTP_STATUS.SERVICE_UNAVAILABLE) {
           debug('uplinks sync failed with timeout error');
-          throw errorUtils.getServiceUnavailable(err.code);
+          throw err;
         }
         // we bubble up the 304 special error case
         if (code === HTTP_STATUS.NOT_MODIFIED) {
@@ -1808,15 +1818,15 @@ class Storage {
     options: Partial<ISyncUplinksOptions>
   ): Promise<Manifest> {
     // we store which uplink is updating the manifest
-    const upLinkMeta = cachedManifest._uplinks[uplink.upname];
+    const upLinkMeta = cachedManifest._uplinks[uplink.uplinkName];
     let _cacheManifest = { ...cachedManifest };
 
-    if (validatioUtils.isObject(upLinkMeta)) {
+    if (validationUtils.isObject(upLinkMeta)) {
       const fetched = upLinkMeta.fetched;
 
       // we check the uplink cache is fresh
       if (fetched && Date.now() - fetched < uplink.maxage) {
-        debug('returning cached manifest for %o', uplink.upname);
+        debug('returning cached manifest for %o', uplink.uplinkName);
         return cachedManifest;
       }
     }
@@ -1832,13 +1842,13 @@ class Storage {
     );
 
     try {
-      _cacheManifest = validatioUtils.normalizeMetadata(_cacheManifest, _cacheManifest.name);
+      _cacheManifest = validationUtils.normalizeMetadata(_cacheManifest, _cacheManifest.name);
     } catch (err: any) {
       this.logger.error({ err }, 'package.json validating error @{!err?.message}\n@{err.stack}');
       throw err;
     }
     // updates the _uplink metadata fields, cache, etc
-    _cacheManifest = updateUpLinkMetadata(uplink.upname, _cacheManifest, etag);
+    _cacheManifest = updateUpLinkMetadata(uplink.uplinkName, _cacheManifest, etag);
     // merge time field cache and remote
     _cacheManifest = mergeUplinkTimeIntoLocalNext(_cacheManifest, remoteManifest);
     // update the _uplinks field in the cache
@@ -2013,7 +2023,7 @@ class Storage {
     for (const up in remoteManifest._uplinks) {
       if (Object.prototype.hasOwnProperty.call(remoteManifest._uplinks, up)) {
         const need_change =
-          !isObject(cacheManifest._uplinks[up]) ||
+          !validationUtils.isObject(cacheManifest._uplinks[up]) ||
           remoteManifest._uplinks[up].etag !== cacheManifest._uplinks[up].etag ||
           remoteManifest._uplinks[up].fetched !== cacheManifest._uplinks[up].fetched;