+++ packages/config/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @verdaccio/config
 
+## 8.0.0-next-8.15
+
+### Patch Changes
+
+- @verdaccio/core@8.0.0-next-8.15
+- @verdaccio/utils@8.1.0-next-8.15
+
+## 8.0.0-next-8.14
+
+### Patch Changes
+
+- b9fea38: chore: cleanup utils
+- bb478f2: chore: expose config utils
+- Updated dependencies [b9fea38]
+  - @verdaccio/core@8.0.0-next-8.14
+  - @verdaccio/utils@8.1.0-next-8.14
+
+## 8.0.0-next-8.13
+
+### Patch Changes
+
+- b3fa5df: fix: uplink processing order
+- e4a1539: chore: package.json maintenance
+- 0607e80: chore: update readme badges and license files
+- Updated dependencies [b3fa5df]
+- Updated dependencies [2eb8cc2]
+- Updated dependencies [e4a1539]
+- Updated dependencies [0607e80]
+  - @verdaccio/utils@8.1.0-next-8.13
+  - @verdaccio/core@8.0.0-next-8.13
+
 ## 8.0.0-next-8.12
 
 ### Patch Changes