+++ packages/core/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @verdaccio/core
 
+## 8.0.0-next-8.15
+
+## 8.0.0-next-8.14
+
+### Patch Changes
+
+- b9fea38: chore: cleanup utils
+
+## 8.0.0-next-8.13
+
+### Patch Changes
+
+- 2eb8cc2: fix: handling for uplink timeouts
+- e4a1539: chore: package.json maintenance
+- 0607e80: chore: update readme badges and license files
+
 ## 8.0.0-next-8.12
 
 ### Patch Changes