+++ packages/core/types/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 13.0.0-next-8.5
+
+### Patch Changes
+
+- 2bcd3ca: chore(config): harmonize configuration options
+
+## 13.0.0-next-8.4
+
+### Patch Changes
+
+- e4a1539: chore: package.json maintenance
+- 0607e80: chore: update readme badges and license files
+
 ## 13.0.0-next-8.3
 
 ### Patch Changes