+++ packages/node-api/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @verdaccio/node-api
 
+## 8.0.0-next-8.15
+
+### Patch Changes
+
+- @verdaccio/server@8.0.0-next-8.15
+- @verdaccio/server-fastify@8.0.0-next-8.15
+- @verdaccio/core@8.0.0-next-8.15
+- @verdaccio/config@8.0.0-next-8.15
+- @verdaccio/logger@8.0.0-next-8.15
+
+## 8.0.0-next-8.14
+
+### Patch Changes
+
+- Updated dependencies [b9fea38]
+- Updated dependencies [bb478f2]
+- Updated dependencies [2bcd3ca]
+- Updated dependencies [8f28186]
+  - @verdaccio/server-fastify@8.0.0-next-8.14
+  - @verdaccio/core@8.0.0-next-8.14
+  - @verdaccio/config@8.0.0-next-8.14
+  - @verdaccio/server@8.0.0-next-8.14
+  - @verdaccio/logger@8.0.0-next-8.14
+
+## 8.0.0-next-8.13
+
+### Patch Changes
+
+- e4a1539: chore: package.json maintenance
+- 0607e80: chore: update readme badges and license files
+- Updated dependencies [b3fa5df]
+- Updated dependencies [2eb8cc2]
+- Updated dependencies [e4a1539]
+- Updated dependencies [0607e80]
+  - @verdaccio/config@8.0.0-next-8.13
+  - @verdaccio/core@8.0.0-next-8.13
+  - @verdaccio/server@8.0.0-next-8.13
+  - @verdaccio/server-fastify@8.0.0-next-8.13
+  - @verdaccio/logger@8.0.0-next-8.13
+
 ## 8.0.0-next-8.12
 
 ### Patch Changes