+++ packages/search-indexer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @verdaccio/search-indexer
 
+## 8.0.0-next-8.4
+
+### Patch Changes
+
+- 8f28186: fix: add legacyMergeConfigs legacy plugins
+
+## 8.0.0-next-8.3
+
+### Patch Changes
+
+- e4a1539: chore: package.json maintenance
+- 0607e80: chore: update readme badges and license files
+
 ## 8.0.0-next-8.2
 
 ### Patch Changes