+++ website/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @verdaccio/website
 
+## 1.0.1-next-8.3
+
+### Patch Changes
+
+- Updated dependencies [2bcd3ca]
+- Updated dependencies [209a650]
+  - @verdaccio/ui-components@4.0.0-next-8.8
+
+## 1.0.1-next-8.2
+
+### Patch Changes
+
+- 0607e80: chore: update readme badges and license files
+- Updated dependencies [ef6864c]
+- Updated dependencies [3763df1]
+- Updated dependencies [c81bd75]
+- Updated dependencies [99f6e11]
+- Updated dependencies [e4a1539]
+- Updated dependencies [4933663]
+- Updated dependencies [0607e80]
+- Updated dependencies [c1faf6d]
+  - @verdaccio/ui-components@4.0.0-next-8.7
+
 ## 1.0.1-next-8.1
 
 ### Patch Changes